# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sortie API
  description: |
    API for Sortie - a centralized application launcher for large organizations.

    Sortie provides a single portal to launch dozens or hundreds of custom applications,
    with support for both URL-based and container-based application launching.
  version: 1.0.0
  contact:
    name: Sortie Team
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://sortie.example.com
    description: Production server
...
security:
  - sessionCookie: []
paths:
...
  /api/sessions:
    get:
      tags:
        - Sessions
      summary: List sessions
      description: Returns a list of active container sessions. Optionally filter by user.
      operationId: listSessions
      parameters:
        - name: user_id
          in: query
          description: Filter sessions by user ID
          schema:
            type: string
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Sessions
      summary: Create a session
      description: |
        Creates a new container session for a container-based application.
        This will provision a Kubernetes pod running the application container
        with a VNC sidecar for browser-based access.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Invalid request or application not configured for container launch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sessions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Session ID
        schema:
          type: string
        example: "sess-abc123"

    get:
      tags:
        - Sessions
      summary: Get a session
      description: Returns details of a specific session including its current status.
      operationId: getSession
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Sessions
      summary: Terminate a session
      description: |
        Terminates an active session. This will delete the associated Kubernetes pod
        and clean up all resources.
      operationId: terminateSession
      responses:
        '204':
          description: Session terminated successfully
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sortie_session
      description: Session cookie set after successful authentication
  schemas:
...
    CreateSessionRequest:
      type: object
      required:
        - app_id
      properties:
        app_id:
          type: string
          description: ID of the application to launch
          example: "vscode"
        user_id:
          type: string
          description: User ID (defaults to "anonymous" if not provided)
          example: "user123"
    SessionResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique session identifier
          example: "sess-abc123"
        user_id:
          type: string
          description: ID of the user who created the session
          example: "user123"
        app_id:
          type: string
          description: ID of the application
          example: "vscode"
        app_name:
          type: string
          description: Name of the application
          example: "Visual Studio Code"
        pod_name:
          type: string
          description: Kubernetes pod name
          example: "sortie-sess-abc123"
        status:
          type: string
          enum:
            - pending
            - creating
            - running
            - terminating
            - terminated
            - failed
          description: Current status of the session
          example: "running"
        websocket_url:
          type: string
          description: WebSocket URL for VNC streaming
          example: "/ws/sessions/sess-abc123"
        created_at:
          type: string
          format: date-time
          description: When the session was created
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: When the session was last updated
          example: "2024-01-15T10:31:00Z"

    # ===================
    # Auth Schemas
    # ===================
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Application not found"
        code:
          type: string
          description: Error code for programmatic handling
          example: "NOT_FOUND"